requirements.txt
# pip install -r requirements.txt
Appium-Python-Client==2.11.0
attrs==23.1.0
browserstack-local==1.2.3
certifi==2023.5.7
charset-normalizer==3.1.0
decorator==5.1.1
docutils==0.20.1
idna==3.4
importlib-resources==5.9.0
jsonpath-ng==1.5.3
jsonschema==4.16.0
kitchen==1.2.6
pkgutil_resolve_name==1.3.10
ply==3.11
psutil==5.9.2
pyrsistent==0.18.1
python-dotenv==0.21.0
requests==2.31.0
robotframework==6.0.2
robotframework-appiumlibrary==2.0.0
robotframework-jsonlibrary==0.5
robotframework-pabot==2.7.0
robotframework-stacktrace==0.4.1
selenium==4.10.0
six==1.16.0
urllib3==2.0.3
zipp==3.8.1
